Corrections, commentaires et formattage

+++ Projet_Bibliotheque/src/java/com/robillard/bibliotheque/controlleur/SupprimerEdition.java
@@ -1,18 +1,13 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
+//Supprime une édition de la bd
+//L'utilisateur doit être connecté et être de type 2 (employé)
+//Le id soumis doit correspondre à une édition dans la base de données
 package com.robillard.bibliotheque.controlleur;
 
 import com.mysql.jdbc.Connection;
 import com.robillard.bibliotheque.modele.classes.Edition;
-import com.robillard.bibliotheque.modele.classes.Ouvrage;
 import com.robillard.bibliotheque.modele.dao.EditionDAO;
-import com.robillard.bibliotheque.modele.dao.OuvrageDAO;
 import com.robillard.bibliotheque.util.Connexion;
 import java.io.IOException;
-import java.io.PrintWriter;
 import java.net.URLEncoder;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -22,56 +17,54 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-/**
- *
- * @author Vengor
- */
-public class SupprimerEdition extends HttpServlet {
+public class SupprimerEdition extends HttpServlet
+{
 
-    /**
-     * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
-     * methods.
-     *
-     * @param request servlet request
-     * @param response servlet response
-     * @throws ServletException if a servlet-specific error occurs
-     * @throws IOException if an I/O error occurs
-     */
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        try
+            throws ServletException, IOException
+    {
+        if (request.getSession().getAttribute("type") != null
+                && Integer.parseInt(request.getSession().getAttribute("type").toString()) >= 2)
         {
-            Class.forName(this.getServletContext().getInitParameter("piloteJDBC"));
-            Connexion.setUrl(this.getServletContext().getInitParameter("urlBd"));
-            Connection cnx = (Connection) Connexion.getInstance();
-            EditionDAO dao = new EditionDAO(cnx);
-            Edition edition = dao.read(request.getParameter("id"));
-            if (edition != null)
+            try
             {
-                dao.delete(edition);
-                String message = "L'" + URLEncoder.encode("é", "UTF-8") + "dition a " 
-                                + URLEncoder.encode("é", "UTF-8") 
-                                + "t" + URLEncoder.encode("é", "UTF-8") + 
-                                " supprim" + URLEncoder.encode("é", "UTF-8") + 
-                                " avec succ" + URLEncoder.encode("è", "UTF-8") + "s";
-                response.sendRedirect("go?action=afficherGestionCatalogue&message="+message
-                                      +"&recherche="+request.getParameter("recherche")
-                                      +"&critere="+request.getParameter("critere"));
+                Class.forName(this.getServletContext().getInitParameter("piloteJDBC"));
+                Connexion.setUrl(this.getServletContext().getInitParameter("urlBd"));
+                Connection cnx = (Connection) Connexion.getInstance();
+                EditionDAO dao = new EditionDAO(cnx);
+                Edition edition = dao.read(request.getParameter("id"));
+                if (edition != null)
+                {
+                    dao.delete(edition);
+                    String message = "L'" + URLEncoder.encode("é", "UTF-8") + "dition a "
+                            + URLEncoder.encode("é", "UTF-8")
+                            + "t" + URLEncoder.encode("é", "UTF-8")
+                            + " supprim" + URLEncoder.encode("é", "UTF-8")
+                            + " avec succ" + URLEncoder.encode("è", "UTF-8") + "s";
+                    response.sendRedirect("go?action=afficherGestionCatalogue&message=" + message
+                            + "&recherche=" + request.getParameter("recherche")
+                            + "&critere=" + request.getParameter("critere"));
+                }
+                else
+                {
+                    RequestDispatcher r = this.getServletContext().getRequestDispatcher("/WEB-INF/gestionCatalogue.jsp");
+                    r.forward(request, response);
+                }
             }
-            else
+            catch (Exception exp)
             {
+                Logger logger = Logger.getLogger("monLogger");
+                logger.log(Level.SEVERE, exp.getMessage());
+                String message = "Une erreur inattendue s'est produite. Veuillez"
+                        + " réessayer plus tard.";
+                request.setAttribute("erreurException", message);
                 RequestDispatcher r = this.getServletContext().getRequestDispatcher("/WEB-INF/gestionCatalogue.jsp");
-                r.forward(request, response);  
+                r.forward(request, response);
             }
         }
-        catch (Exception exp)
+        else
         {
-            Logger logger = Logger.getLogger("monLogger");
-            logger.log(Level.SEVERE, exp.getMessage());
-            String message = "Une erreur inattendue s'est produite. Veuillez"
-                    + " réessayer plus tard.";
-            request.setAttribute("erreurException", message);
-            RequestDispatcher r = this.getServletContext().getRequestDispatcher("/WEB-INF/gestionCatalogue.jsp");
+            RequestDispatcher r = this.getServletContext().getRequestDispatcher("/index.jsp");
             r.forward(request, response);
         }
     }
@@ -87,7 +80,8 @@ public class SupprimerEdition extends HttpServlet {
      */
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
+            throws ServletException, IOException
+    {
         processRequest(request, response);
     }
 
@@ -101,7 +95,8 @@ public class SupprimerEdition extends HttpServlet {
      */
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
+            throws ServletException, IOException
+    {
         processRequest(request, response);
     }
 
@@ -111,7 +106,8 @@ public class SupprimerEdition extends HttpServlet {
      * @return a String containing servlet description
      */
     @Override
-    public String getServletInfo() {
+    public String getServletInfo()
+    {
         return "Short description";
     }// </editor-fold>
