+++ Projet_Bibliotheque/src/java/com/robillard/bibliotheque/controlleur/SupprimerExemplaire.java
@@ -1,13 +1,13 @@
+//Supprime un exemplaire de la bd
+//L'utilisateur doit être connecté et être de type 2 (employé)
+//Le id soumis doit correspondre à un exemplaire dans la base de données
 package com.robillard.bibliotheque.controlleur;
 
 import com.mysql.jdbc.Connection;
-import com.robillard.bibliotheque.modele.classes.Edition;
 import com.robillard.bibliotheque.modele.classes.Exemplaire;
-import com.robillard.bibliotheque.modele.dao.EditionDAO;
 import com.robillard.bibliotheque.modele.dao.ExemplaireDAO;
 import com.robillard.bibliotheque.util.Connexion;
 import java.io.IOException;
-import java.io.PrintWriter;
 import java.net.URLEncoder;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -17,41 +17,52 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class SupprimerExemplaire extends HttpServlet {
+public class SupprimerExemplaire extends HttpServlet
+{
 
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        try
+            throws ServletException, IOException
+    {
+        if (request.getSession().getAttribute("type") != null
+                && Integer.parseInt(request.getSession().getAttribute("type").toString()) >= 2)
         {
-            Class.forName(this.getServletContext().getInitParameter("piloteJDBC"));
-            Connexion.setUrl(this.getServletContext().getInitParameter("urlBd"));
-            Connection cnx = (Connection) Connexion.getInstance();
-            ExemplaireDAO dao = new ExemplaireDAO(cnx);
-            Exemplaire exemplaire = dao.read(request.getParameter("id"));
-            if (exemplaire != null)
+            try
             {
-                dao.delete(exemplaire);
-                String message = "L'exemplaire a " 
-                                + URLEncoder.encode("é", "UTF-8") 
-                                + "t" + URLEncoder.encode("é", "UTF-8") + 
-                                " supprim" + URLEncoder.encode("é", "UTF-8") + 
-                                " avec succ" + URLEncoder.encode("è", "UTF-8") + "s";
-                response.sendRedirect("go?action=afficherModificationEdition&message="+message+"&id="+exemplaire.getEdition().getId());
+                Class.forName(this.getServletContext().getInitParameter("piloteJDBC"));
+                Connexion.setUrl(this.getServletContext().getInitParameter("urlBd"));
+                Connection cnx = (Connection) Connexion.getInstance();
+                ExemplaireDAO dao = new ExemplaireDAO(cnx);
+                Exemplaire exemplaire = dao.read(request.getParameter("id"));
+                if (exemplaire != null)
+                {
+                    dao.delete(exemplaire);
+                    String message = "L'exemplaire a "
+                            + URLEncoder.encode("é", "UTF-8")
+                            + "t" + URLEncoder.encode("é", "UTF-8")
+                            + " supprim" + URLEncoder.encode("é", "UTF-8")
+                            + " avec succ" + URLEncoder.encode("è", "UTF-8") + "s";
+                    response.sendRedirect("go?action=afficherModificationEdition&message=" + message + "&id=" + exemplaire.getEdition().getId());
+                }
+                else
+                {
+                    RequestDispatcher r = this.getServletContext().getRequestDispatcher("/WEB-INF/gestionCatalogue.jsp");
+                    r.forward(request, response);
+                }
             }
-            else
+            catch (Exception exp)
             {
+                Logger logger = Logger.getLogger("monLogger");
+                logger.log(Level.SEVERE, exp.getMessage());
+                String message = "Une erreur inattendue s'est produite. Veuillez"
+                        + " réessayer plus tard.";
+                request.setAttribute("erreurException", message);
                 RequestDispatcher r = this.getServletContext().getRequestDispatcher("/WEB-INF/gestionCatalogue.jsp");
-                r.forward(request, response);  
+                r.forward(request, response);
             }
         }
-        catch (Exception exp)
+        else
         {
-            Logger logger = Logger.getLogger("monLogger");
-            logger.log(Level.SEVERE, exp.getMessage());
-            String message = "Une erreur inattendue s'est produite. Veuillez"
-                    + " réessayer plus tard.";
-            request.setAttribute("erreurException", message);
-            RequestDispatcher r = this.getServletContext().getRequestDispatcher("/WEB-INF/gestionCatalogue.jsp");
+            RequestDispatcher r = this.getServletContext().getRequestDispatcher("/index.jsp");
             r.forward(request, response);
         }
     }
@@ -67,7 +78,8 @@ public class SupprimerExemplaire extends HttpServlet {
      */
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
+            throws ServletException, IOException
+    {
         processRequest(request, response);
     }
 
@@ -81,7 +93,8 @@ public class SupprimerExemplaire extends HttpServlet {
      */
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
+            throws ServletException, IOException
+    {
         processRequest(request, response);
     }
 
@@ -91,7 +104,8 @@ public class SupprimerExemplaire extends HttpServlet {
      * @return a String containing servlet description
      */
     @Override
-    public String getServletInfo() {
+    public String getServletInfo()
+    {
         return "Short description";
     }// </editor-fold>
